# Use proper units when printing beam time