**Before:**
```
   0.00s:                 from   1 ->   1 actions 1024 1024 1024
  18.37s:    409654.60 us from   6 ->   6 actions 1024  256 1024    4
  29.96s:    125520.07 us from  12 ->  11 actions  256  256 1024    4    4
  35.11s:     86910.17 us from  12 ->  12 actions  128  256 1024    4    4    2
  40.31s:     86910.17 us from  12 ->  12 actions  128  256 1024    4    4    2
*** LLVM      10 r_128_256_1024_4_4_2                      arg  3 mem  0.01 GB tm     81.16ms/    93.54ms (    26.46 GFLOPS    0.2|19.9    GB/s) ['matmul']
```
**After:**
```
   0.00s:                from   1 ->   1 actions 1024 1024 1024
  22.33s:       423.46ms from   6 ->   6 actions 1024  256 1024    4
  34.47s:       132.41ms from  12 ->  11 actions  256  256 1024    4    4
  39.71s:        90.93ms from  12 ->  12 actions  128  256 1024    4    4    2
  44.87s:        90.93ms from  12 ->  12 actions  128  256 1024    4    4    2
*** LLVM      10 r_128_256_1024_4_4_2                      arg  3 mem  0.01 GB tm     78.43ms/    90.03ms (    27.38 GFLOPS    0.2|20.6    GB/s) ['matmul']
```