@geohot 2025-02-17
You should refactor DEBUG=2 in this PR